Change size_t to long in OpenMP loops

+++ psi4/src/psi4/libmints/matrix.cc
@@ -1198,7 +1198,7 @@ void Matrix::apply_denominator(const Matrix *const plus) {
             lhs = matrix_[h][0];
             rhs = plus->matrix_[h][0];
 #pragma omp parallel for simd
-            for (size_t ij = 0; ij < size; ++ij) {
+            for (long ij = 0; ij < size; ++ij) {
                 lhs[ij] /= rhs[ij];
             }
         }
